Rename speech optimization modules
